+++ test/test-ocsp.hs
@@ -1,9 +1,10 @@
-{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RecordWildCards, OverloadedStrings #-}
 
 module Main where
 
 import Data.X509.AIA
 import Data.X509.OCSP
+import Data.X509.Validation
 import Data.X509
 import Data.PEM
 import Test.HUnit
@@ -14,7 +15,6 @@ import Data.ASN1.Types
 import Data.ASN1.Encoding
 import Data.ASN1.BinaryEncoding
 import Data.List (uncons)
-import Control.Monad
  
 toCertificate :: ByteString -> Certificate
 toCertificate cert = either error getCertificate $ do
@@ -38,13 +38,24 @@ testOCSPRequest :: Certificate -> Certificate -> ByteString -> Test
 testOCSPRequest cert issuerCert = TestCase . (buildRequest @?=) . L.fromStrict
     where buildRequest = fst $ encodeOCSPRequest cert issuerCert
  
-testOCSPResponse :: CertId -> ByteString -> Test
-testOCSPResponse certId resp =
+testOCSPResponse :: Maybe OCSPResponse -> Test
+testOCSPResponse resp =
     TestCase $ getRespStatus @?= Just OCSPRespCertGood
-    where getRespStatus = either (const Nothing)
-              (fmap ocspRespPayload >=>
-                  fmap (ocspRespCertStatus . ocspRespCertData)
-              ) $ decodeOCSPResponse certId $ L.fromStrict resp
+    where getRespStatus = ocspRespCertStatus . ocspRespCertData <$>
+              (resp >>= ocspRespPayload)
+ 
+verifyOCSPResponse :: Certificate -> Maybe OCSPResponse -> Test
+verifyOCSPResponse issuerCert resp =
+    TestCase $ getVerificationStatus @?= Just SignaturePass
+    where getVerificationStatus = (`verifySignature'` issuerCert) <$> resp
+
+verifySignature' :: OCSPResponse -> Certificate -> SignatureVerification
+verifySignature' resp Certificate {..}
+    | Just OCSPResponseVerificationData {..} <-
+        getOCSPResponseVerificationData resp
+    , Right (alg, _) <- fromASN1 ocspRespSignatureAlg =
+        verifySignature alg certPubKey ocspRespDer ocspRespSignature
+    | otherwise = SignatureFailed SignatureInvalid
 
 main :: IO ()
 main = do
@@ -54,14 +65,18 @@ main = do
     let certId = snd $ encodeOCSPRequestASN1 certS certI
 
     reqDer <- B.readFile "test/data/req.der"
-    let req = either (error . show) id $ decodeASN1 DER $ L.fromStrict reqDer
+    let req = showError $ decodeASN1' DER reqDer
 
     respDer <- B.readFile "test/data/resp.der"
+    let resp = showError $ decodeOCSPResponse certId $ L.fromStrict respDer
 
     runTestTTAndExit $ TestList
         [TestLabel "testAIA"             $ testAIA certS
         ,TestLabel "testOCSPRequestASN1" $ testOCSPRequestASN1 certS certI req
         ,TestLabel "testOCSPRequest"     $ testOCSPRequest certS certI reqDer
-        ,TestLabel "testOCSPResponse"    $ testOCSPResponse certId respDer
+        ,TestLabel "testOCSPResponse"    $ testOCSPResponse resp
+        ,TestLabel "verifyOCSPResponse"  $ verifyOCSPResponse certI resp
         ]
+
+    where showError = either (error . show) id